+++ rumqttc/src/mqttbytes/v4/mod.rs
@@ -1,4 +1,5 @@
 use super::*;
+use core::fmt;
 
 mod connack;
 mod connect;
@@ -27,6 +28,113 @@ pub use suback::*;
 pub use subscribe::*;
 pub use unsuback::*;
 pub use unsubscribe::*;
+
+/// Packet type from a byte
+///
+/// ```ignore
+///          7                          3                          0
+///          +--------------------------+--------------------------+
+/// byte 1   | MQTT Control Packet Type | Flags for each type      |
+///          +--------------------------+--------------------------+
+///          |         Remaining Bytes Len  (1/2/3/4 bytes)        |
+///          +-----------------------------------------------------+
+///
+/// <https://docs.oasis-open.org/mqtt/mqtt/v3.1.1/os/mqtt-v3.1.1-os.html#_Toc385349207>
+/// ```
+#[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd)]
+pub struct FixedHeader {
+    /// First byte of the stream. Used to identify packet types and
+    /// several flags
+    pub byte1: u8,
+    /// Length of fixed header. Byte 1 + (1..4) bytes. So fixed header
+    /// len can vary from 2 bytes to 5 bytes
+    /// 1..4 bytes are variable length encoded to represent remaining length
+    pub fixed_header_len: usize,
+    /// Remaining length of the packet. Doesn't include fixed header bytes
+    /// Represents variable header + payload size
+    pub remaining_len: usize,
+}
+
+impl FixedHeader {
+    pub fn new(byte1: u8, remaining_len_len: usize, remaining_len: usize) -> FixedHeader {
+        FixedHeader {
+            byte1,
+            fixed_header_len: remaining_len_len + 1,
+            remaining_len,
+        }
+    }
+
+    pub fn packet_type(&self) -> Result<PacketType, Error> {
+        let num = self.byte1 >> 4;
+        match num {
+            1 => Ok(PacketType::Connect),
+            2 => Ok(PacketType::ConnAck),
+            3 => Ok(PacketType::Publish),
+            4 => Ok(PacketType::PubAck),
+            5 => Ok(PacketType::PubRec),
+            6 => Ok(PacketType::PubRel),
+            7 => Ok(PacketType::PubComp),
+            8 => Ok(PacketType::Subscribe),
+            9 => Ok(PacketType::SubAck),
+            10 => Ok(PacketType::Unsubscribe),
+            11 => Ok(PacketType::UnsubAck),
+            12 => Ok(PacketType::PingReq),
+            13 => Ok(PacketType::PingResp),
+            14 => Ok(PacketType::Disconnect),
+            _ => Err(Error::InvalidPacketType(num)),
+        }
+    }
+
+    /// Returns the size of full packet (fixed header + variable header + payload)
+    /// Fixed header is enough to get the size of a frame in the stream
+    pub fn frame_length(&self) -> usize {
+        self.fixed_header_len + self.remaining_len
+    }
+}
+
+/// Checks if the stream has enough bytes to frame a packet and returns fixed header
+/// only if a packet can be framed with existing bytes in the `stream`.
+/// The passed stream doesn't modify parent stream's cursor. If this function
+/// returned an error, next `check` on the same parent stream is forced start
+/// with cursor at 0 again (Iter is owned. Only Iter's cursor is changed internally)
+pub fn check(stream: Iter<u8>, max_packet_size: usize) -> Result<FixedHeader, Error> {
+    // Create fixed header if there are enough bytes in the stream
+    // to frame full packet
+    let stream_len = stream.len();
+    let fixed_header = parse_fixed_header(stream)?;
+
+    // Don't let rogue connections attack with huge payloads.
+    // Disconnect them before reading all that data
+    if fixed_header.remaining_len > max_packet_size {
+        return Err(Error::PayloadSizeLimitExceeded {
+            pkt_size: fixed_header.remaining_len,
+            max: max_packet_size,
+        });
+    }
+
+    // If the current call fails due to insufficient bytes in the stream,
+    // after calculating remaining length, we extend the stream
+    let frame_length = fixed_header.frame_length();
+    if stream_len < frame_length {
+        return Err(Error::InsufficientBytes(frame_length - stream_len));
+    }
+
+    Ok(fixed_header)
+}
+
+/// Parses fixed header
+fn parse_fixed_header(mut stream: Iter<u8>) -> Result<FixedHeader, Error> {
+    // At least 2 bytes are necessary to frame a packet
+    let stream_len = stream.len();
+    if stream_len < 2 {
+        return Err(Error::InsufficientBytes(2 - stream_len));
+    }
+
+    let byte1 = stream.next().unwrap();
+    let (len_len, len) = length(stream)?;
+
+    Ok(FixedHeader::new(*byte1, len_len, len))
+}
 
 /// Encapsulates all MQTT packet types
 #[derive(Debug, Clone, PartialEq, Eq)]